Fix some issues that came to light only after asking for comments from FCC community!
2 bugs found when starting in the upper left diagonal and the comp will play center. Then user goes to the bottom left corner and the comp will block which creates a possible win for comp which the user then block. This will form a 'T' on it's side.
At this point the computer did not play since `takeOppositeCorner()` was returning 2 possible answers and forming a 'a3c3' square and not choosing one of them. This lead the user to continue play and capture a win. However game play did not stop after the user win. So the comp won as well.
That second bug was fixed by added a line to the Main.js component so that the `compTurn()` would not happen when there was a found result from `findResult()`.
The first bug got fix by bring `takeOppositeCorner()` inline with `sidesEmpty()` and `cornersEmpty()` with finding the length and getting a random number with lodash to choose one or the other.

I fix the test to handle this case by making sure the square that is used in the action call to dispatch the store, was part of the original game board!

add last commit to notes to add to blog post